Use _aligned_malloc on Windows.
On 32-bit Windows, malloc doesn't give us memory with sufficiently
strong alignment for AVX vectors.

Note that Windows also has a family of allocators that provide guarantee
exact misalignment.  Using these would simplify the ivar layout logic
slightly, because we wouldn't need to account for the refcount pointer.
It's not clear that it's actually worth doing though, because that logic
already exists.