Fix some weak pointer bugs.
These most likely led to some intermittent failures in weak references.
The `objc_moveWeak` and `objc_copyWeak` functions were assuming that the
target may contain valid data, but the spec for these says that it may
be uninitialised data.  We were reading uninitialised data and then
seeing if it happened to be a valid object.  Most of the time it was
likely to be nil, so this wasn't an easily reproduceable problem...

I believe this is the correct fix for PR #95.

Fixes #95