Windows: Support building with strict compatibility
This commit adds the `STRICT_APPLE_COMPATIBILITY` CMake option which will force `BOOL` to defined as an `signed char` on Windows instead of an `int`, as having different types for `BOOL` on different platforms creates serialization issues.

On MSYS, the `BOOL` type is defined in` minwindef` like this:

```
  typedef int BOOL;
```

To prevent this header from defining a conflicting `BOOL` type, this commit also defines `__OBJC_BOOL`.