Persist STRICT_APPLE_COMPATIBILITY in config.h
On Windows, the value of `STRICT_APPLE_COMPATIBILITY` determines whether `BOOL` is a char or an int.  This value must be consistent across all libraries that use libobjc2.  Rather than requiring all libraries to explicitly define `STRICT_APPLE_COMPATIBILITY`, persist the value of `STRICT_APPLE_COMPATIBILITY` in `objc-config.h` and include that in `runtime.h`.