Support defining C99 bool as the BOOL type
Whether to use C99 bool can be set at build time using
-DOBJC_BOOL_IS_BOOL_MODE=auto/always/never.

Default is to use C99 bool if __OBJC_BOOL_IS_BOOL is defined and
set to 1, which is currently the case for Clang on newer Apple
platforms. This matches the behavior of Apple's libobjc (especially
when combined with STRICT_APPLE_COMPATIBILITY).

(See
https://www.jviotti.com/2024/01/05/is-objective-c-bool-a-boolean-type-it-depends.html
for more information about the __OBJC_BOOL_IS_BOOL definition.)