Fix leak of weak references
In the case where an object has been deallocated, fetching the old value of a weak reference to it returns nil.  If we are assigning nil to the weak reference we can't abort processing early (when we check that old and new values are the same) as that will leave the weak reference in place when the calling code thinks it has destroyed it.